+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## Unreleased
+
+### Features
+
+* refonte de la page supervision avec statuts ping, SSH, modules actifs et fraicheur explicite
+* correction du refresh manuel de supervision pour retester aussi SSH
+
 ## [0.28.0](https://github.com/grandpurs45/my-server-manager/compare/v0.27.0...v0.28.0) (2026-06-09)